pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "api_demo_server"
authors = [{name = "Maksim Beliaev", email = "beliaev.m.s@gmail.com"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "psycopg2-binary"
]

# development dependencies
[project.optional-dependencies]
test = [
    "pre-commit",
    "mypy",
    "pytest",
    "pytest-cov",
]

deploy = [
    "flit==3.7.1",
]